#47 navcontroller 수정
#47 navcontroller 수정
