+++ src/main/webapp/WEB-INF/views/mypage/myblacklist.jsp
@@ -35,7 +35,7 @@ a {
   <script>
     $(document).ready( function() {
       $("#topnav").load("topnav.do");
-      $("#leftmenu").load("mynav.do");
+      $("#leftMenu").load("mynav.do");
     });
   </script>
 	<nav id="topnav"></nav>
