+++ src/main/webapp/WEB-INF/views/mypage/mymain.jsp
@@ -66,7 +66,7 @@ body{
   <script>
     $(document).ready( function() {
       $("#topnav").load("topnav.do");
-      $("#leftmenu").load("mynav.do");
+      $("#leftMenu").load("mynav.do");
     });
   </script>
 	<nav id="topnav"></nav>
