+++ src/main/webapp/WEB-INF/views/mypage/mywrite.jsp
@@ -37,7 +37,7 @@ a {
   <script>
     $(document).ready( function() {
       $("#topnav").load("topnav.do");
-      $("#leftmenu").load("mynav.do");
+      $("#leftMenu").load("mynav.do");
     });
   </script>
 	<nav id="topnav"></nav>
